Add Sentence Parser with default \r delimiter
Add IPacket and IParse interfaces which together can be used to create
generic packet parsers.
The first implementation of these interfaces is that of a sentence parser,
namely packets which are delimited by a single or couple of characters.
The default delimiter is \r

IPacket contains a single read method, getString which will return the
parsed string if it has been completed.
IParse has a single Parse method which is used by the caller to add new
incoming data to the packet and an onPacket call back which is called on
completion of a packet.